refactor(agent): extract the import_elements applier
applyOps.ts sits exactly at the 399-line limit that check:max-lines enforces, so
anything added to it fails the check. import_elements is the most self-contained
applier, so it moves out first to leave headroom.

Pure move: the function body is unchanged, and it takes a narrow scene interface
rather than importing the Scene class so the two modules stay decoupled. The
existing tests cover it.
